+++ src/backend_new/main.py
@@ -9,12 +9,9 @@ from collections.abc import Generator
 from typing import Any
 
 # HELPER LIBRARIES
-from backend_new.utils.helper_funcs import (read_json_file, write_json_file,
-                                            write_config,
-                                            questionary_select, questionary_checkbox,
-                                            load_env_file,
-                                            clear_temp_dir,
-                                            write_env_key)
+# from backend_new.utils.helper_funcs import (questionary_select, questionary_checkbox)
+from backend_new.utils.functions.filesystem import read_json_file, write_json_file, write_config, write_env_key, \
+    load_env_file, clear_temp_dir
 
 from backend_new.core.workflow import WorkflowManager
