--- src/backend_new/utils/helper_funcs.py
@@ -1,253 +1,0 @@
-from typing import Any
-from pathlib import Path
-from questionary import Choice
-import questionary as q
-from itertools import batched
-from dotenv import load_dotenv, set_key
-import json
-import base58
-import os
-import re
-import gdown
-from collections.abc import Generator
-
-# CONSTANTS
-from backend_new.utils.constants import ENV_FILE, TEMP_DIR, DEFAULT_ENV_VARS, CONFIG_FILE, DEFAULT_DICTS_FOLDER_LINK, DICTS_DIR, DEFAULT_DICTS
-
-# LOGGER
-from backend_new.utils.logger import Logger
-logger = Logger(__name__)
-
-# region questionary
-def questionary_select(question_to_ask: str,
-                       choose_data: list[dict | str] | list[Choice],
-                       use_shortcuts: bool = True,
-                       enable_pages: bool = False,
-                       batch_data: bool = False,
-                       batch_size: int = 5,
-                       enable_all: str = '',
-                       enable_exit: str = '',
-                       extra_navigation_options: list[dict | str] | list[Choice] = None) -> Any:
-    """
-    Abstraction layer for questionary select
-    :param question_to_ask: Prompt to display
-    :param choose_data: The select data containing a list of dicts or Choice objects
-    :param use_shortcuts: Uses shortcuts for each option, can be customized
-    :param enable_pages: Enables the page navigation buttons
-    :param batch_data: Splits data into batches, enable_pages must be enabled to work
-    :param batch_size: How many items per page
-    :param enable_all: Enables the "All" button
-    :param enable_exit: Enables the "Exit" button
-    :param extra_navigation_options: Extra navigation options to add to the end of the list
-    :return: The selected option
-    """
-
-    if batch_data and not enable_pages:
-        raise Exception("Cannot batch data without enabling pages")
-
-    choose_data = [option for option in choose_data]
-
-    def _add_navigation_buttons(input_list: list[dict | str]):
-        if enable_pages or enable_all != '' or enable_exit != '' or extra_navigation_options:
-            input_list.append(q.Separator())
-        if enable_pages:
-            input_list.append(Choice("Next ->", value = "__next__", shortcut_key='w'))
-            input_list.append(Choice("Previous <-", value= "__prev__", shortcut_key='s'))
-        if enable_all != '':
-            input_list.append(Choice(enable_all, value= "__all__", shortcut_key='a'))
-        if enable_exit != '':
-            input_list.append(Choice(enable_exit, value= "__exit__", shortcut_key='e'))
-        if extra_navigation_options:
-            for option in extra_navigation_options:
-                input_list.append(option)
-
-    if enable_pages and batch_data:
-        offset = 0
-
-        while True:
-            batched_choose_data = list(batched(choose_data, batch_size))
-            selected_batched_choose_data = list(batched_choose_data[offset])
-            _add_navigation_buttons(selected_batched_choose_data)
-
-            user_choice = q.select(question_to_ask, choices=selected_batched_choose_data, use_shortcuts=use_shortcuts).ask()
-            if user_choice == "__next__":
-                if offset + 1 >= len(batched_choose_data):
-                    pass
-                else:
-                    offset += 1
-            elif user_choice == "__prev__":
-                if offset > 0:
-                    offset -= 1
-            elif user_choice == "__all__":
-                break
-            elif user_choice == "__exit__":
-                break
-            else:
-                break
-    else:
-        _add_navigation_buttons(choose_data)
-        user_choice = q.select(question_to_ask, choices=choose_data, use_shortcuts=use_shortcuts).ask()
-
-    return user_choice
-
-def questionary_checkbox(question_to_ask: str,
-                         choice_data : list[dict | str] | list[Choice]) -> Any:
-    """
-    Abstraction layer for questionary checkbox
-    :param question_to_ask: Prompt to display
-    :param choice_data: a list of dicts or Choice objects
-    :return: The selected options
-    """
-    return q.checkbox(question_to_ask, choices=choice_data).ask()
-# endregion
-
-# region json read and write
-def read_json_file(file_path: Path, safe_fail: bool = False) -> dict:
-    """
-    Reads a JSON file in a directory
-    :param file_path: File path to the JSON file
-    :param safe_fail: If true, when file doesn't exist, doesn't raise an exception
-    :return: A dict containing the JSON data
-    """
-    if not file_path.exists():
-        logger.warning(f"File {file_path} does not exist")
-        if safe_fail:
-            return {}
-        else:
-            raise FileNotFoundError(f"File {file_path} does not exist")
-
-    try:
-        return json.loads(file_path.read_text(encoding="utf-8"))
-    except json.JSONDecodeError as e:
-        raise json.JSONDecodeError(f"The provided JSON file is not a valid JSON file or is corrupted, please remove / repair this file. {file_path}", e.doc, e.pos)
-
-def write_json_file(file_path: Path, payload: Any, keys: list[str] = None) -> None:
-    """
-    Writes data to a JSON file in a directory
-    :param file_path: File path to the JSON file
-    :param payload: Data to write
-    :param keys: A list of keys to navigate the JSON file
-    """
-    data = read_json_file(file_path)
-
-    current_level = data
-    for key in keys[:-1]:
-        if key not in current_level or not isinstance(current_level[key], dict):
-            current_level[key] = {}
-        current_level = current_level[key]
-
-    if keys:
-        current_level[keys[-1]] = payload
-
-    file_path.write_text(json.dumps(data, indent=4))
-
-# specialized reader
-def read_config() -> dict[str, Any]:
-    """
-    Reads the config file data
-    :return: Config file data
-    :rtype: dict[str, Any]
-    """
-    return read_json_file(CONFIG_FILE)
-
-def write_config(payload: Any, keys: list[str] = None) -> None:
-    """
-    Writes to the config file
-    :param payload: Data to write
-    :param keys: A list of keys to navigate the JSON file
-    """
-    write_json_file(CONFIG_FILE, payload, keys)
-# endregion
-
-# region env keys
-def write_env_key(value: str, key: str) -> None:
-    set_key(dotenv_path=ENV_FILE, key_to_set=key, value_to_set=value, quote_mode="never")
-# end region
-
-# region base 58 conv
-def str_to_base58(string: str) -> str:
-    """
-    Encodes a string to Base58
-    :param string: Input string to encode
-    :return: Encoded string in Base58 format
-    """
-    return base58.b58encode(string.encode('utf-8')).decode('utf-8')
-
-def base58_to_str(string: str) -> str:
-    """
-    Decodes a Base58-encoded string to normal text
-    :param string: Base58-encoded string to decode
-    :return: Normal text string
-    """
-    return base58.b58decode(string).decode('utf-8')
-# endregion
-
-# region file system stuff
-def load_env_file(safe_empty: bool = True) -> dict[Any, str | None]:
-    """
-    Loads the environment variables from the .env file.
-    Creates a new .env file if it doesn't exist.
-    DOES NOT check if the file is filled
-    """
-    if ENV_FILE.exists():
-        load_dotenv(dotenv_path=ENV_FILE)
-        logger.debug("Loaded .env file.")
-    else:
-        if safe_empty:
-            pass
-        else:
-            logger.critical("No .env file found. Creating empty .env file. Do NOT reorder the variables")
-            ENV_FILE.write_text("\n".join([f"{var}=" for var in DEFAULT_ENV_VARS]))
-            raise FileNotFoundError(".env file not found, creating one. Please add your credentials to the .env file.")
-
-    load_dotenv()
-    return dict([(var, os.getenv(var))for var in DEFAULT_ENV_VARS])
-
-def clear_temp_dir() -> None:
-    if q.confirm(f"Are you sure you want to clear the .temp directory?\n{TEMP_DIR}", auto_enter=False, default=False).ask():
-        for item in TEMP_DIR.iterdir():
-            if item.suffix == ".wav":
-                item.unlink()
-            elif item.suffix == ".json":
-                item.unlink()
-            else:
-                logger.info(f"{item} is not a a .wav or .json file")
-    else:
-        logger.info("Cancelling operation")
-# endregion
-
-# region japanese specific
-
-# This checks for:
-# - Hiragana: \u3040-\u309f
-# - Katakana: \u30a0-\u30ff
-# - Kanji (CJK Unified Ideographs): \u4e00-\u9faf
-JAPANESE_CHAR_PATTERN = re.compile(r"[\u3040-\u309f\u30a0-\u30ff\u4e00-\u9faf]")
-
-def contains_japanese(string: str) -> bool:
-    if type(string) is not str:
-        return False
-
-    return bool(JAPANESE_CHAR_PATTERN.search(string))
-# endregion
-
-# region download gdrive
-def download_google_drive(link: str, output_path: Path) -> None:
-    gdown.download(link, str(output_path), quiet=False)
-# endregion
-
-# region download dicts
-def download_all_dicts() -> Generator[str, None, None]:
-    try:
-        files = gdown.download_folder(DEFAULT_DICTS_FOLDER_LINK, skip_download=True)
-        re_dicts_search = [re.compile(f".*{item}.*") for item in DEFAULT_DICTS]
-
-        for file_info in files:
-            if any([complied.match(file_info.path) for complied in re_dicts_search]):
-                yield f"Downloading: {file_info.path}"
-                gdown.download(id=file_info.id, output=DICTS_DIR, use_cookies=False)
-                yield f"Finished Downloading: {file_info.path}"
-    except Exception as e:
-        raise e
-
-# endregion
